geminipredictions/pyproject.toml
[project]
name = "geminipredictions-mcp"
version = "0.1.0"
description = "Gemini Predictions market data API as an MCP server -- powered by Invariant Protocol"
requires-python = ">=3.10"
dependencies = [
    "invariant-protocol @ git+https://github.com/jim-technologies/invariantprotocol.git#subdirectory=python",
    "httpx>=0.27",
    "pyyaml>=6.0.3",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/geminipredictions_mcp"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
exclude = [
    "src/geminipredictions_mcp/gen",
]
